Move pause/resume to abstract Kafka connector, fix config bug
