+++ datastream-kafka-connector/src/main/java/com/linkedin/datastream/connectors/kafka/AbstractKafkaConnector.java
@@ -4,17 +4,25 @@ import java.time.Duration;
 import java.time.temporal.ChronoUnit;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Properties;
+import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.atomic.AtomicInteger;
 
+import org.apache.kafka.clients.consumer.Consumer;
+import org.apache.kafka.common.PartitionInfo;
 import org.slf4j.Logger;
 
 import com.linkedin.datastream.common.Datastream;
+import com.linkedin.datastream.common.DatastreamMetadataConstants;
 import com.linkedin.datastream.common.DatastreamRuntimeException;
+import com.linkedin.datastream.common.DatastreamSource;
+import com.linkedin.datastream.common.DatastreamUtils;
+import com.linkedin.datastream.common.JsonUtils;
 import com.linkedin.datastream.common.ReflectionUtils;
 import com.linkedin.datastream.common.ThreadUtils;
 import com.linkedin.datastream.common.VerifiableProperties;
@@ -133,7 +141,69 @@ public abstract class AbstractKafkaConnector implements Connector {
   @Override
   public void validateUpdateDatastreams(List<Datastream> datastreams, List<Datastream> allDatastreams)
       throws DatastreamValidationException {
-    // nop
+    // validate for paused partitions
+    validatePausedPartitions(datastreams, allDatastreams);
+  }
+
+
+  private void validatePausedPartitions(List<Datastream> datastreams, List<Datastream> allDatastreams)
+      throws DatastreamValidationException {
+    for (Datastream ds : datastreams) {
+      validatePausedPartitions(ds, allDatastreams);
+    }
+  }
+
+  private void validatePausedPartitions(Datastream datastream, List<Datastream> allDatastreams)
+      throws DatastreamValidationException {
+    Map<String, Set<String>> pausedSourcePartitionsMap = DatastreamUtils.getDatastreamSourcePartitions(datastream);
+
+    for (String source : pausedSourcePartitionsMap.keySet()) {
+      Set<String> newPartitions = pausedSourcePartitionsMap.get(source);
+
+      // Validate that partitions actually exist and convert any "*" to actual list of partitions.
+      // For that, get the list of existing partitions first.
+      List<PartitionInfo> partitionInfos = getKafkaTopicPartitions(datastream, source);
+      Set<String> allPartitions = new HashSet<>();
+      for (PartitionInfo info : partitionInfos) {
+        allPartitions.add(String.valueOf(info.partition()));
+      }
+
+      // if there is any * in the new list, just convert it to actual list of partitions.
+      if (newPartitions.contains(DatastreamMetadataConstants.REGEX_PAUSE_ALL_PARTITIONS_IN_A_TOPIC)) {
+        newPartitions.clear();
+        newPartitions.addAll(allPartitions);
+      } else {
+        // Else make sure there aren't any partitions that don't exist.
+        newPartitions.retainAll(allPartitions);
+      }
+    }
+
+    // Now write back the set to datastream
+    datastream.getMetadata()
+        .put(DatastreamMetadataConstants.PAUSED_SOURCE_PARTITIONS_KEY, JsonUtils.toJson(pausedSourcePartitionsMap));
+  }
+
+  private List<PartitionInfo> getKafkaTopicPartitions(Datastream datastream, String topic)
+      throws DatastreamValidationException {
+    List<PartitionInfo> partitionInfos = null;
+
+    DatastreamSource source = datastream.getSource();
+    String connectionString = source.getConnectionString();
+
+    KafkaConnectionString parsed = KafkaConnectionString.valueOf(connectionString);
+
+    try (Consumer<?, ?> consumer = KafkaConnectorTask.createConsumer(_consumerFactory, _consumerProps,
+        "KafkaConnectorPartitionFinder", parsed)) {
+      partitionInfos = consumer.partitionsFor(topic);
+      if (partitionInfos == null) {
+        throw new DatastreamValidationException("Can't get partition info from kafka for topic: " + topic);
+      }
+    } catch (Exception e) {
+      throw new DatastreamValidationException(
+          "Exception received while retrieving info on kafka topic partitions: " + e);
+    }
+
+    return partitionInfos;
   }
 
 }
