+++ datastream-kafka-connector/src/main/java/com/linkedin/datastream/connectors/kafka/mirrormaker/FlushlessKafkaMirrorMakerConnectorTask.java
@@ -45,7 +45,6 @@ class FlushlessKafkaMirrorMakerConnectorTask extends KafkaMirrorMakerConnectorTa
         new KafkaMirrorMakerCheckpoint(datastreamProducerRecord.getCheckpoint());
     _flushlessProducer.send(datastreamProducerRecord, sourceCheckpoint.getTopic(), sourceCheckpoint.getPartition(),
         sourceCheckpoint.getOffset());
-
   }
 
   @Override
